requirements.txt
aiogram

python-dotenv
